more refactors and styling to watch detail page
